+++ 試驗/訓練過渡格式/test聽拍單元試驗.py
@@ -5,7 +5,6 @@ from unittest.case import skip
 from django.test import TestCase
 
 
-from 試驗.加資料.加資料試驗 import 加資料試驗
 from 臺灣言語資料庫.資料模型 import 聽拍表
 
 
@@ -37,7 +36,7 @@ class 聽拍資料試驗(TestCase):
             }
         })
 
-    def test_加詞(self):
+    def test_一句一句話(self):
         self.資料 = 聽拍表._加資料(self.詞內容)
         self.assertEqual(json.loads(self.資料.聽拍資料), [
             {'語者': '阿宏', '內容': 'li1', '開始時間': 0.0, '結束時間': 1.2},
@@ -78,6 +77,30 @@ class 聽拍資料試驗(TestCase):
         self.assertEqual(self.資料表.objects.all().count(), 0)
 
     def test_聽拍資料無內容欄位(self):
+        self.詞內容['聽拍資料'] = [
+            {'語者': '阿宏', '開始時間': 0.0, '結束時間': 1.2},
+            {'語者': '阿莉', '開始時間': 1.2, '結束時間': 2.0},
+        ]
+        self.assertRaises(KeyError, super(加聽拍資料試驗, self).test_加詞)
+        self.assertEqual(self.資料表.objects.all().count(), 0)
+
+    def test_無語者欄位(self):
+        self.詞內容['聽拍資料'] = [
+            {'語者': '阿宏', '開始時間': 0.0, '結束時間': 1.2},
+            {'語者': '阿莉', '開始時間': 1.2, '結束時間': 2.0},
+        ]
+        self.assertRaises(KeyError, super(加聽拍資料試驗, self).test_加詞)
+        self.assertEqual(self.資料表.objects.all().count(), 0)
+
+    def test_無開始時間欄位(self):
+        self.詞內容['聽拍資料'] = [
+            {'語者': '阿宏', '開始時間': 0.0, '結束時間': 1.2},
+            {'語者': '阿莉', '開始時間': 1.2, '結束時間': 2.0},
+        ]
+        self.assertRaises(KeyError, super(加聽拍資料試驗, self).test_加詞)
+        self.assertEqual(self.資料表.objects.all().count(), 0)
+
+    def test_聽拍資料無結束時間欄位(self):
         self.詞內容['聽拍資料'] = [
             {'語者': '阿宏', '開始時間': 0.0, '結束時間': 1.2},
             {'語者': '阿莉', '開始時間': 1.2, '結束時間': 2.0},
